requirements.txt
asgiref==3.11.1
Django==5.2.4
django-storages==1.14.4
boto3==1.35.93
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-allauth==65.15.1
dj-rest-auth[with_social]==7.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
google-auth-oauthlib==1.3.1
pillow==10.3.0
PyJWT==2.12.1
razorpay==2.0.1
sqlparse==0.5.5
psycopg2-binary==2.9.10
